Rename Group_09_report__1_ (5).ipynb to Data Science and Visualization/Group_09_report__1_.ipynb
This project explores the use of machine learning models to predict the outcome of loan applications based on economic and personal data. The dataset consists of various features regarding the applicant's financial background and loan details. I have applied both decision tree and random forest models to analyze and predict loan approvals. The goal is to identify the most influential factors affecting lending decisions and to assess the accuracy of predictive models in financial risk management.